Add context_annotations file, improve coding style
