+++ src/sparc/curation/tools/annotations/scaffold.py
@@ -3,53 +3,6 @@ import os
 from sparc.curation.tools.definitions import MIMETYPE_TO_FILETYPE_MAP, MIMETYPE_TO_PARENT_FILETYPE_MAP, MIMETYPE_TO_CHILDREN_FILETYPE_MAP, FILE_LOCATION_COLUMN, MANIFEST_DIR_COLUMN, FILENAME_COLUMN, ADDITIONAL_TYPES_COLUMN, SOURCE_OF_COLUMN, \
     DERIVED_FROM_COLUMN
 
-
-class ScaffoldAnnotationError(object):
-
-    def __init__(self, message, location, mime):
-        self._message = message
-        self._location = location
-        self._mime = mime
-
-    def get_location(self):
-        return self._location
-
-    def get_error_message(self):
-        return f'Error: {self._message}'
-
-    def get_mime(self):
-        return self._mime
-
-    # def __str__(self):
-    #     return f'Error: {self._message}'
-
-
-class NotAnnotatedError(ScaffoldAnnotationError):
-    def __init__(self, location, mime):
-        fileType = MIMETYPE_TO_FILETYPE_MAP.get(mime, 'unknown')
-        message = f"Found Scaffold '{fileType}' file that is not annotated '{location}'."
-        super(NotAnnotatedError, self).__init__(message, location, mime)
-
-class IncorrectSourceOfError(ScaffoldAnnotationError):
-    def __init__(self, location, mime):
-        fileType = MIMETYPE_TO_FILETYPE_MAP.get(mime, 'unknown')
-        childrenFileType = MIMETYPE_TO_CHILDREN_FILETYPE_MAP.get(mime, 'unknown')
-        message = f"Found '{fileType}' file '{location}' either has no {childrenFileType} file or it's annotated to an incorrect file."
-        super(IncorrectSourceOfError, self).__init__(message, location, mime)
-
-class IncorrectDerivedFromError(ScaffoldAnnotationError):
-    def __init__(self, location, mime):
-        fileType = MIMETYPE_TO_FILETYPE_MAP.get(mime, 'unknown')
-        parentFileType = MIMETYPE_TO_PARENT_FILETYPE_MAP.get(mime, 'unknown')
-        message = f"Found '{fileType}' file '{location}' either has no derived from file or it's not derived from a scaffold '{parentFileType}' file."
-        super(IncorrectDerivedFromError, self).__init__(message, location, mime)
-
-class IncorrectAnnotationError(ScaffoldAnnotationError):
-    def __init__(self, location, mime):
-        fileType = MIMETYPE_TO_FILETYPE_MAP.get(mime, 'unknown')
-        message = f"File '{location}' either does not exist or is not a scaffold '{fileType}' file."
-        super(IncorrectAnnotationError, self).__init__(message, location, mime)
-
 class ScaffoldAnnotation(object):
     """
     TODO use this class to wrap one dataframe row to an object.
